+++ domain/src/main/kotlin/es/pedrazamiguez/expenseshareapp/domain/service/addon/ExactAddOnResolver.kt
@@ -11,6 +11,7 @@ import java.math.RoundingMode
  */
 class ExactAddOnResolver : AddOnAmountResolver {
 
+    @Suppress("UNUSED_PARAMETER")
     override fun resolve(
         normalizedInput: BigDecimal,
         decimalDigits: Int,
