+++ features/balances/src/main/kotlin/es/pedrazamiguez/expenseshareapp/features/balance/presentation/viewmodel/handler/WithdrawalFeeHandler.kt
@@ -1,7 +1,7 @@
 package es.pedrazamiguez.expenseshareapp.features.balance.presentation.viewmodel.handler
 
 import es.pedrazamiguez.expenseshareapp.core.designsystem.presentation.formatter.FormattingHelper
-import es.pedrazamiguez.expenseshareapp.domain.service.ExpenseCalculatorService
+import es.pedrazamiguez.expenseshareapp.domain.service.ExchangeRateCalculationService
 import es.pedrazamiguez.expenseshareapp.domain.usecase.currency.GetExchangeRateUseCase
 import es.pedrazamiguez.expenseshareapp.features.balance.presentation.mapper.AddCashWithdrawalUiMapper
 import es.pedrazamiguez.expenseshareapp.features.balance.presentation.viewmodel.action.AddCashWithdrawalUiAction
@@ -20,7 +20,7 @@ import timber.log.Timber
  */
 class WithdrawalFeeHandler(
     private val getExchangeRateUseCase: GetExchangeRateUseCase,
-    private val expenseCalculatorService: ExpenseCalculatorService,
+    private val exchangeRateCalculationService: ExchangeRateCalculationService,
     private val mapper: AddCashWithdrawalUiMapper,
     private val formattingHelper: FormattingHelper
 ) : AddCashWithdrawalEventHandler {
@@ -130,7 +130,7 @@ class WithdrawalFeeHandler(
 
         val sourceDecimalPlaces = state.feeCurrency?.decimalDigits ?: 2
         val targetDecimalPlaces = state.groupCurrency?.decimalDigits ?: 2
-        val calculatedConverted = expenseCalculatorService.calculateGroupAmountFromDisplayRate(
+        val calculatedConverted = exchangeRateCalculationService.calculateGroupAmountFromDisplayRate(
             sourceAmountString = state.feeAmount,
             displayRateString = state.feeExchangeRate,
             sourceDecimalPlaces = sourceDecimalPlaces,
@@ -149,7 +149,7 @@ class WithdrawalFeeHandler(
         if (!state.showFeeExchangeRateSection) return
 
         val sourceDecimalPlaces = state.feeCurrency?.decimalDigits ?: 2
-        val impliedRate = expenseCalculatorService.calculateImpliedDisplayRateFromStrings(
+        val impliedRate = exchangeRateCalculationService.calculateImpliedDisplayRateFromStrings(
             sourceAmountString = state.feeAmount,
             groupAmountString = state.feeConvertedAmount,
             sourceDecimalPlaces = sourceDecimalPlaces
