+++ features/expenses/src/test/kotlin/es/pedrazamiguez/expenseshareapp/features/expense/presentation/viewmodel/handler/SubmitEventHandlerTest.kt
@@ -8,6 +8,7 @@ import es.pedrazamiguez.expenseshareapp.domain.enums.SplitType
 import es.pedrazamiguez.expenseshareapp.domain.model.AddOn
 import es.pedrazamiguez.expenseshareapp.domain.model.Expense
 import es.pedrazamiguez.expenseshareapp.domain.model.ExpenseSplit
+import es.pedrazamiguez.expenseshareapp.domain.service.AddOnCalculationService
 import es.pedrazamiguez.expenseshareapp.domain.service.ExpenseCalculatorService
 import es.pedrazamiguez.expenseshareapp.domain.service.ExpenseValidationService
 import es.pedrazamiguez.expenseshareapp.domain.service.RemainderDistributionService
@@ -62,6 +63,7 @@ class SubmitEventHandlerTest {
         handler = SubmitEventHandler(
             addExpenseUseCase = mockk(relaxed = true),
             expenseValidationService = ExpenseValidationService(splitCalculatorFactory),
+            addOnCalculationService = AddOnCalculationService(),
             expenseCalculatorService = ExpenseCalculatorService(),
             remainderDistributionService = RemainderDistributionService(),
             setGroupLastUsedCurrencyUseCase = mockk(relaxed = true),
